requirements.txt
python-docx==1.1.0
num2words==0.5.12
mysql-connector-python==8.2.0
